# docs/swagger.yaml
definitions:
  models.APIResponse:
    properties:
      data:
        description: Optional data payload
      error:
        description: |-
          Optional error
          example: Something went wrong
        type: string
      message:
        description: |-
          The response message
          example: Success
        type: string
    type: object
...
  models.Department:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Engineering
        type: string
      updated_at:
        type: string
    type: object
  models.Software:
    properties:
      created_at:
        type: string
      description:
        example: Communication Suite
        type: string
      id:
        example: 1
        type: integer
      name:
        example: ClickUp
        type: string
      type:
        description: e.g., License, SaaS, etc.
        example: SaaS
        type: string
      updated_at:
        type: string
    type: object
...
  models.Staff:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/models.Department'
      department_id:
        example: 2
        type: integer
      email:
        example: john.doe@shuttlers.co
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      status:
        example: Active
        type: string
      team:
        $ref: '#/definitions/models.Team'
      team_id:
        example: 5
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Team:
    properties:
      created_at:
        type: string
      department:
        $ref: '#/definitions/models.Department'
      department_id:
        example: 2
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Frontend
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/software:
    get:
      parameters:
      - description: Search by software name
        in: query
        name: search
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Software'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get paginated software with filtering and search
      tags:
      - Software
    post:
      consumes:
      - application/json
      parameters:
      - description: Software to create
        in: body
        name: software
        required: true
        schema:
          $ref: '#/definitions/models.Software'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Software'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create new software
      tags:
      - Software
...
  /api/software/{id}:
    delete:
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete software and revoke all its assignments
      tags:
      - Software
    get:
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Software'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get a software item by ID
      tags:
      - Software
    put:
      consumes:
      - application/json
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      - description: Software fields to update
        in: body
        name: software
        required: true
        schema:
          $ref: '#/definitions/models.Software'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Software'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update existing software
      tags:
      - Software
  /api/software/{software_id}/assigned-staff:
    get:
      parameters:
      - description: Software ID
        in: path
        name: id
        required: true
        type: integer
      - description: Search by staff name
        in: query
        name: search
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Staff'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get list of staff assigned to a specific software
      tags:
      - Software
  /api/software/plain:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Software'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get all software
      tags:
      - Software
...
swagger: "2.0"